controltia/requirements.txt
absl-py==0.5.0
arrow==0.14.7
astor==0.7.1
atomicwrites==1.3.0
attrs==19.3.0
bleach==1.5.0
bokeh==0.12.14
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.3
convert2php==0.0.1
CVXcanon==0.1.1
cvxpy==0.4.11
cx-Oracle==7.0.0
cycler==0.10.0
decorator==4.3.0
dill==0.2.8.2
Django==2.2
django-multi-email-field==0.6.1
ecos==2.0.7rc1
et-xmlfile==1.0.1
fastcache==1.0.2
gast==0.2.0
geographiclib==1.49
geopy==1.17.0
grpcio==1.15.0
h5py==2.8.0
html5lib==0.9999999
ics==0.7
idna==2.7
importlib-metadata==1.6.0
ipython-genutils==0.2.0
jdcal==1.4
Jinja2==2.10
joblib==0.12.0
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.1.6
Markdown==3.0
MarkupSafe==1.0
matplotlib==2.1.2
more-itertools==8.2.0
multiprocess==0.70.5
mysql-connector-python==8.0.18
nbformat==4.4.0
networkx==2.1
numpy==1.13.3
openpyxl==2.5.0
ortools==6.7.4973
packaging==17.1
pandas==0.22.0
pathlib2==2.3.5
pd==0.0.1
Pillow==7.1.2
plotly==2.7.0
pluggy==0.13.1
protobuf==3.6.1
psycopg2==2.8.5
PuLP==1.6.8
py==1.8.1
PyMySQL==0.9.3
pyodbc==4.0.21
pyparsing==2.2.1
PypeR==1.1.2
pyreadline==2.1
pytest==5.4.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.13
qrcode==6.1
requests==2.19.1
retrying==1.3.3
scikit-learn==0.19.1
scipy==1.0.0
scs==2.0.2
selenium==3.141.0
six==1.11.0
sklearn==0.0
sqlparse==0.3.1
TatSu==4.4.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
virtualenv==16.7.2
walmart-calendar==1.1.0
wcwidth==0.1.9
Werkzeug==0.14.1
xlrd==1.1.0
zipp==1.2.0
